A4A: Address review feedback on the referrals data layer
Group referrals by client in queryFn so the transform also applies when
the query is preloaded, and move the non-GET endpoints to mutators.ts.
