+++ packages/api-core/src/agency-referrals/fetchers.ts
@@ -16,25 +16,3 @@ export async function fetchReferralCommissionPayout(
 		apiNamespace: 'wpcom/v2',
 	} );
 }
-
-export async function archiveReferral(
-	agencyId: number,
-	referralId: number
-): Promise< { success: boolean } > {
-	return wpcom.req.post( {
-		path: `/agency/${ agencyId }/referrals/${ referralId }/archive`,
-		apiNamespace: 'wpcom/v2',
-		method: 'PUT',
-	} );
-}
-
-export async function resendReferralEmail(
-	agencyId: number,
-	referralId: number
-): Promise< { success: boolean } > {
-	return wpcom.req.post( {
-		path: `/agency/${ agencyId }/referrals/${ referralId }/resend`,
-		apiNamespace: 'wpcom/v2',
-		method: 'POST',
-	} );
-}
