+++ packages/api-core/src/agency-referrals/mutators.ts
@@ -1,0 +1,23 @@
+import { wpcom } from '../wpcom-fetcher';
+
+export async function archiveReferral(
+	agencyId: number,
+	referralId: number
+): Promise< { success: boolean } > {
+	return wpcom.req.post( {
+		path: `/agency/${ agencyId }/referrals/${ referralId }/archive`,
+		apiNamespace: 'wpcom/v2',
+		method: 'PUT',
+	} );
+}
+
+export async function resendReferralEmail(
+	agencyId: number,
+	referralId: number
+): Promise< { success: boolean } > {
+	return wpcom.req.post( {
+		path: `/agency/${ agencyId }/referrals/${ referralId }/resend`,
+		apiNamespace: 'wpcom/v2',
+		method: 'POST',
+	} );
+}
